use cosign to verify the release artifact

+++ ct.sh
@@ -88,7 +88,19 @@ install_chart_testing() {
         mkdir -p "$cache_dir"
 
         echo "Installing chart-testing ${version}..."
+        CT_CERT=https://github.com/helm/chart-testing/releases/download/$version/chart-testing_${version#v}_linux_amd64.tar.gz.pem
+        CT_SIG=https://github.com/helm/chart-testing/releases/download/$version/chart-testing_${version#v}_linux_amd64.tar.gz.sig
+
         curl --retry 5 --retry-delay 1 -sSLo ct.tar.gz "https://github.com/helm/chart-testing/releases/download/$version/chart-testing_${version#v}_linux_amd64.tar.gz"
+        cosign verify-blob --certificate $CT_CERT --signature $CT_SIG \
+          --certificate-identity "https://github.com/helm/chart-testing/.github/workflows/release.yaml@refs/tags/${version#v}" \
+          --certificate-oidc-issuer "https://token.actions.githubusercontent.com" "chart-testing_${version#v}_linux_amd64.tar.gz"
+        retVal=$?
+        if [[ "$retVal" -ne 0 ]]; then
+          log_error "Unable to validate chart-testing version: ${version}"
+          exit 1
+        fi
+
         tar -xzf ct.tar.gz -C "$cache_dir"
         rm -f ct.tar.gz
